+++ packages/terra-clinical-item-display/CHANGELOG.md
@@ -7,7 +7,7 @@
   * `span` tag around non-strong style text.
 
 * Changed
-  * Make screen reader support for text styling more robust by using terra-visually-hidden-text.
+  * Improved screen reader support for text styling by utilizing terra-visually-hidden-text.
 
 ## 4.7.0 - (May 15, 2023)