+++ packages/terra-clinical-item-display/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Added
-  * a11ylabel prop to ItemComment icon.
+  * Added `a11ylabel` prop to ItemComment icon.
 
 * Added
   * `strong` tag around strong style text.