+++ packages/terra-clinical-item-display/CHANGELOG.md
@@ -4,10 +4,9 @@
 
 * Added
   * Added `a11ylabel` prop to ItemComment icon.
-
-* Added
   * `strong` tag around strong style text.
   * `span` tag around non-strong style text.
+  * Added translations for hidden text styling.
 
 * Changed
   * Improved screen reader support for text styling by utilizing terra-visually-hidden-text.